Module Xstringext.String

include module type of Stdlib.String

val length : string -> int

val get : string -> int -> char

val set : bytes -> int -> char -> unit

val create : int -> bytes

val make : int -> char -> string

val init : int -> (int -> char) -> string

val copy : string -> string

val sub : string -> int -> int -> string

val fill : bytes -> int -> int -> char -> unit

val blit : string -> int -> bytes -> int -> int -> unit

val concat : string -> string list -> string

val iter : (char -> unit) -> string -> unit

val iteri : (int -> char -> unit) -> string -> unit

val map : (char -> char) -> string -> string

val mapi : (int -> char -> char) -> string -> string

val trim : string -> string

val escaped : string -> string

val index : string -> char -> int

val index_opt : string -> char -> int option

val rindex : string -> char -> int

val rindex_opt : string -> char -> int option

val index_from : string -> int -> char -> int

val index_from_opt : string -> int -> char -> int option

val rindex_from : string -> int -> char -> int

val rindex_from_opt : string -> int -> char -> int option

val contains : string -> char -> bool

val contains_from : string -> int -> char -> bool

val rcontains_from : string -> int -> char -> bool

val uppercase : string -> string

val lowercase : string -> string

val capitalize : string -> string

val uncapitalize : string -> string

val uppercase_ascii : string -> string

val lowercase_ascii : string -> string

val capitalize_ascii : string -> string

val uncapitalize_ascii : string -> string

type t = string

val compare : t -> t -> int

val equal : t -> t -> bool

val split_on_char : char -> string -> string list

val to_seq : t -> char Stdlib.Seq.t

val to_seqi : t -> (int * char) Stdlib.Seq.t

val of_seq : char Stdlib.Seq.t -> t

val unsafe_get : string -> int -> char

val unsafe_set : bytes -> int -> char -> unit

val unsafe_blit : string -> int -> bytes -> int -> int -> unit

val unsafe_fill : bytes -> int -> int -> char -> unit

val of_char : char -> string

val rev_map : (char -> char) -> string -> string

Map a string to a string, applying the given function in reverse order.

val rev_iter : (char -> unit) -> string -> unit

Iterate over the characters in a string in reverse order.

val fold_left : ('a -> char -> 'a) -> 'a -> string -> 'a

Fold over the characters in a string.

val fold_right : (char -> 'a -> 'a) -> string -> 'a -> 'a

Iterate over the characters in a string in reverse order.

val explode : string -> char list

Split a string into a list of characters.

val implode : char list -> string

Concatenate a list of characters into a string.

val endswith : string -> string -> bool

True if string 'x' ends with suffix 'suffix'

val startswith : string -> string -> bool

True if string 'x' starts with prefix 'prefix'

val isspace : char -> bool

True if the character is whitespace

val strip : (char -> bool) -> string -> string

Removes all the characters from the ends of a string for which the predicate is true

val escaped : ?⁠rules:(char * string) list -> string -> string

Backward-compatible string escaping, defaulting to the built-in OCaml string escaping but allowing an arbitrary mapping from characters to strings.

val split_f : (char -> bool) -> string -> string list

Take a predicate and a string, return a list of strings separated by runs of characters where the predicate was true

val split : ?⁠limit:int -> char -> string -> string list

split a string on a single char

val rtrim : string -> string

FIXME document me|remove me if similar to strip

val has_substr : string -> string -> bool

True if sub is a substr of str

val find_all : string -> string -> int list

find all occurences of needle in haystack and return all their respective index

val replace : string -> string -> string -> string

replace all f substring in s by t

val filter_chars : string -> (char -> bool) -> string

filter chars from a string

val map_unlikely : string -> (char -> string option) -> string

map a string trying to fill the buffer by chunk

val sub_to_end : string -> int -> string

a substring from the specified position to the end of the string

val sub_before : char -> string -> string

a substring from the start of the string to the first occurrence of a given character, excluding the character

val sub_after : char -> string -> string

a substring from the first occurrence of a given character to the end of the string, excluding the character